backend/sandbox/backend/image_processing.py
# Handles image-related tasks, so includes AI model calls and image processing - 


# Implement a function to load and preprocess the image file with Llama 3.2-Vision
    # Include logic to interact with Llama Vision 
# https://huggingface.co/meta-llama/Llama-3.2-11B-Vision-Instruct
    # Make sure to update your transformers installation via pip install --upgrade transformers.
        # pip install --upgrade transformers

import requests
import torch
from PIL import Image
from transformers import MllamaForConditionalGeneration, AutoProcessor

model_id = "meta-llama/Llama-3.2-11B-Vision-Instruct"

model = MllamaForConditionalGeneration.from_pretrained(
    model_id,
    torch_dtype=torch.bfloat16,
    device_map="auto",
)
processor = AutoProcessor.from_pretrained(model_id)

url = "https://huggingface.co/datasets/huggingface/documentation-images/resolve/0052a70beed5bf71b92610a43a52df6d286cd5f3/diffusers/rabbit.jpg"
image = Image.open(requests.get(url, stream=True).raw)

messages = [
    {"role": "user", "content": [
        {"type": "image"},
        {"type": "text", "text": "If I had to write a haiku for this one, it would be: "}
    ]}
]
input_text = processor.apply_chat_template(messages, add_generation_prompt=True)
inputs = processor(
    image,
    input_text,
    add_special_tokens=False,
    return_tensors="pt"
).to(model.device)

output = model.generate(**inputs, max_new_tokens=30)
print(processor.decode(output[0]))

# Ensure proper error handling
# add error handling for unsupported formats or corrupted files
# Use libraries like Pillow (PIL) or OpenCV for image processing

# Fine-tune image processing as needed for Llama 3.2-Vision input requirements...
# ...biological (medical) images may need specific preprocessing steps.
# Refer to Llama 3.2-Vision documentation for image input specifications.
# https://www.youtube.com/watch?v=co0lDx5J23o