requirements.txt
# Requirements for Wetlands-prototype-MVP
# This file lists all the Python packages required to run the Wetlands-prototype-MVP application.

fastapi  # FastAPI is a modern web framework for building APIs with Python
uvicorn  # Uvicorn is an ASGI server for running FastAPI applications
python-multipart  # Enables handling of multipart/form-data, useful for file uploads
pillow  # Pillow is a library for image processing in Python
requests  # Requests is a library for making HTTP requests in Python
python-dotenv  # Python-dotenv loads environment variables from a .env file
huggingface_hub==1.2.3  # Hugging Face Hub is used to access models and datasets from Hugging Face

# Phase 1 MVP = one narrow pipeline: Local image → Backend → AI API → Structured suggestions → Output
# So we don't need heavy ML libraries or databases at this stage

# requirements.txt: If new libraries like Hugging Face's transformers or ollama are required, they should be added to this file.